+++ app/src/main/java/com/etsdk/app/huov7/model/AddressList.java
@@ -19,14 +19,14 @@ public class AddressList extends BaseModel {
     }
 
     public static class DataBean{
-        private int count;
+        private double count;
         private List<AddressInfo> list;
 
-        public int getCount() {
+        public double getCount() {
             return count;
         }
 
-        public void setCount(int count) {
+        public void setCount(double count) {
             this.count = count;
         }
