+++ app/src/main/java/com/etsdk/app/huov7/model/CouponBeanList.java
@@ -18,14 +18,14 @@ public class CouponBeanList extends BaseModel{
     }
 
     public static class DataBean {
-        private int count;
+        private double count;
         private List<CouponListItem> list;
 
-        public int getCount() {
+        public double getCount() {
             return count;
         }
 
-        public void setCount(int count) {
+        public void setCount(double count) {
             this.count = count;
         }
